Add chat button to map.
